Created Mock DB, added login unit tests
